// fron/cliente/src/app/models/Edm05.ts
export class Edm05titulo{
    _id?: number;
    edm05titulo1: string;
    edm05titulo2: string;
    edm05titulo3: string;
    edm05titulo4: string;
    edm05titulo5: string; 
    edm05titulo6: string;
    edm05titulo7: string;
    edm05titulo8: string;
    edm05titulo9: string;
    edm05titulo10: string;
    edm05titulo11: string;
    edm05titulo12: string;
    edm05titulo13: string;
    edm05titulo14: string;
    edm05titulo15:string;
    edm05titulo16: string;
    edm05titulo17: string;
    edm05titulo18: string;
    edm05titulo19: string;
    edm05titulo20: string;
    edm05titulo21: string;
    edm05titulo22: string;
    edm05titulo23: string;
    edm05titulo24: string;
    edm05titulo25: string;
    edm05titulo26: string;
    

    constructor(edm05titulo1:string,edm05titulo2:string,edm05titulo3:string,edm05titulo4:string ,edm05titulo5: string,
       edm05titulo6: string, edm05titulo7: string, edm05titulo8: string, edm05titulo9: string, edm05titulo10: string, 
       edm05titulo11: string, edm05titulo12: string, edm05titulo13: string, edm05titulo14: string, edm05titulo15: string,
       edm05titulo16: string, edm05titulo17: string, edm05titulo18: string, edm05titulo19: string, edm05titulo20: string,
       edm05titulo21: string, edm05titulo22: string, edm05titulo23: string, edm05titulo24: string, edm05titulo25: string,
       edm05titulo26: string,
       ){
        this.edm05titulo1=edm05titulo1;
        this.edm05titulo2=edm05titulo2;
        this.edm05titulo3=edm05titulo3;
        this.edm05titulo4=edm05titulo4;
        this.edm05titulo5=edm05titulo5;
        this.edm05titulo6=edm05titulo6;
        this.edm05titulo7=edm05titulo7;
        this.edm05titulo8=edm05titulo8;
        this.edm05titulo9=edm05titulo9;
        this.edm05titulo10=edm05titulo10;
        this.edm05titulo11=edm05titulo11;
        this.edm05titulo12=edm05titulo12;
        this.edm05titulo13=edm05titulo13;
        this.edm05titulo14=edm05titulo14;
        this.edm05titulo15=edm05titulo15;
        this.edm05titulo16=edm05titulo16;
        this.edm05titulo17=edm05titulo17;
        this.edm05titulo18=edm05titulo18;
        this.edm05titulo19=edm05titulo19;
        this.edm05titulo20=edm05titulo20;
        this.edm05titulo21=edm05titulo21;
        this.edm05titulo22=edm05titulo22;
        this.edm05titulo23=edm05titulo23;
        this.edm05titulo24=edm05titulo24;
        this.edm05titulo25=edm05titulo25;
        this.edm05titulo26=edm05titulo26;

    }

}